requirement.txt
click==7.1.2
dlib==19.21.0
Flask==1.1.2
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
opencv-contrib-python==3.4.6.27
opencv-python==3.4.6.27
picamera==1.13
Pillow==8.0.1
scipy==1.5.4
Werkzeug==1.0.1
